Cythonize terminal.
